added wind text to measurand text hierarchy
