+++ lib/wrf_forecast/text.rb
@@ -2,11 +2,12 @@
 # @Author: Benjamin Held
 # @Date:   2020-05-06 18:11:59
 # @Last Modified by:   Benjamin Held
-# @Last Modified time: 2020-05-06 18:18:19
+# @Last Modified time: 2020-05-06 18:47:28
 
 module WrfForecast
 
   require "wrf_forecast/forecast/measurand_text"
   require "wrf_forecast/forecast/english/temperature_text"
+  require "wrf_forecast/forecast/english/wind_text"
 
 end
